Feat: useProfileImage 커스텀로직 분리
